fix(server): keep websocket heartbeat live during request dispatch
Request handlers ran inline in the connection loop's select arm, so a
long JSON-RPC call (agent turn, blocked store query) delayed Pong
servicing and heartbeat ticks. When the handler returned, tokio's random
select among ready branches could observe pong_pending still set and kill
a healthy connection.

Move dispatch onto a serialized per-connection worker task fed by a
bounded channel: the connection loop only forwards frames, services
Pongs, and ticks the heartbeat. One worker preserves FIFO response
ordering; a full backlog now rejects with a JSON-RPC error instead of
growing memory without bound.

Also extract websocket test support (AppState builder + sandbox-tolerant
client helpers) into websocket_test_support so the connection-loop tests
and the new dispatch-worker tests share one fixture without pushing
websocket.rs past its size limit.

Closes #1984
